Update notifier::SendAudioRaw to use default type from decoder

+++ include/view/base/notifier.h
@@ -52,7 +52,7 @@ class Notifier {
    * @param buffer Audio samples
    * @param buff_size Sample count
    */
-  virtual void SendAudioRaw(int* buffer, int buffer_size) = 0;
+  virtual void SendAudioRaw(uint8_t* buffer, int buffer_size) = 0;
 
   /**
    * @brief Notify UI with error code from some background operation
